requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0,<3.0
pydantic>=2.4
accelerate>=0.34.0
safetensors>=0.4.3
psycopg2-binary==2.9.10